Grouped the args so they're now more easily read and imported
