+++ docs/backends/FadeCandy.md
@@ -3,4 +3,4 @@ To use the Fadecandy backend, please ensure that you are running the Fadecandy s
 A fork of the Fadecandy repo can be found [here](https://github.com/TheMariday/fadecandy)
 
 Use 
-`--backend fadecandy --server 127.0.0.1:7890` to enable this backend, adjusting the server IP and port where needed
+`marimapper fadecandy --server 127.0.0.1:7890` to enable this backend, adjusting the server IP and port where needed
